CafuModelT: Corrected the datatype of some members from "int" to "unsigned int".
git-svn-id: https://srv7.svn-repos.de/dev123/projects/cafu/trunk@213 5fe0b3f3-48df-4727-b460-e6dda9226989

+++ Libs/Models/Loader_lwo.cpp
@@ -345,7 +345,7 @@ void LoaderLwoT::Load(ArrayT<CafuModelT::JointT>& Joints, ArrayT<CafuModelT::Mes
                         const unsigned long               MeshVertexNr=WeightVerts[PolyVert.index][Nr];
                         const CafuModelT::MeshT::VertexT& MeshVertex  =Mesh.Vertices[MeshVertexNr];
 
-                        assert(MeshVertex.FirstWeightIdx==PolyVert.index);
+                        assert(MeshVertex.FirstWeightIdx==(unsigned int)PolyVert.index);
                         assert(MeshVertex.NumWeights==1);
 
                         if (MeshVertex.u==TexCoord.x && MeshVertex.v==TexCoord.y)
